requirements.txt
absl-py==1.2.0
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
auto-tune @ file:///root/selfgz177322614/compiler/lib64/auto_tune-0.1.0-py3-none-any.whl
Babel==2.10.3
backcall==0.2.0
backports.lzma==0.0.14
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
easydict==1.10
entrypoints==0.4
executing==1.1.1
face-alignment==1.4.1
facenet-pytorch==2.5.3
fastjsonschema==2.16.2
filelock==3.13.1
fonttools==4.37.4
frozenlist==1.4.1
fsspec==2024.2.0
future==0.18.2
gast==0.5.3
google-auth==2.12.0
google-auth-oauthlib==0.4.6
grpcio==1.49.1
h5py==3.7.0
hccl @ file:///root/selfgz177322614/compiler/lib64/hccl-0.1.0-py3-none-any.whl
huggingface-hub==0.21.4
idna==3.4
imageio==2.34.1
importlib-metadata==5.0.0
importlib_resources==6.4.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.21.0
jupyter_client==7.4.2
jupyterlab==3.4.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.16.0
kiwisolver==1.4.4
lazy_loader==0.4
libcst==1.2.0
llvmlite==0.42.0
lmdb==1.3.0
loguru==0.7.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.8.4
matplotlib-inline==0.1.6
mindspore-ascend @ file:///tmp/mindspore_ascend-1.8.0-cp39-cp39-linux_aarch64.whl
mistune==2.0.4
mmcv==1.2.7
mpmath==1.2.1
multidict==6.0.5
mypy-extensions==1.0.0
nbclassic==0.4.6
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.2.1
ninja==1.11.1.1
notebook==6.4.12
notebook-shim==0.1.0
numba==0.59.1
numpy==1.26.4
oauthlib==3.2.1
opc-tool @ file:///root/selfgz177322614/compiler/lib64/opc_tool-0.1.0-py3-none-any.whl
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prettytable==3.10.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.5
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest-runner==6.0.0
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
regex==2023.12.25
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
schedule-search @ file:///root/selfgz177322614/compiler/lib64/schedule_search-0.1.0-py3-none-any.whl
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.13.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.1
sympy==1.11.1
te @ file:///root/selfgz177322614/compiler/lib64/te-0.4.0-py3-none-any.whl
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.16.0
threadpoolctl==3.1.0
tifffile==2024.4.18
tinycss2==1.1.1
tokenizers==0.15.2
tomli==2.0.1
topi @ file:///root/selfgz177322614/compiler/lib64/topi-0.4.0-py3-none-any.whl
torch @ file:///home/qinjinghui/code/torch-2.1.0-cp39-cp39-manylinux_2_17_aarch64.manylinux2014_aarch64.whl#sha256=de7d63c6ecece118684415a3dbd4805af4a4c1ee1490cccf7405d8c240a481b4
torch-npu==2.1.0
torch_geometric==2.3.1
torchvision==0.16.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transformers==4.38.2
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
yapf==0.32.0
yarl==1.9.4
zipp==3.9.0
